new api_root url to list accessible endpoints. Moving shared api policy(permission,throttle) in core app
