+++ movie/api_urls.py
@@ -1,12 +1,11 @@
 
-app_name = "movie"
-
 from django.urls import path, include
 from . import views 
 
+app_name = "movie"
 
 # including the api route here
 urlpatterns = [
-    path('list/', views.MovieListView.as_view()),
-    path('detail/<int:pk>/', views.MovieDetailView.as_view()),
+    path('list/', views.MovieListView.as_view(), name='list'),
+    path('detail/<int:pk>/', views.MovieDetailView.as_view(), name='detail'),
 ]
